# docs/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: Cybertech E-Commerce API
  description: |
    Backend REST API for the Cybertech e-commerce platform.

    Covers the full e-commerce lifecycle: product catalog, search,
    cart, wishlist, orders, payments (Stripe), reviews and moderation,
    bank cards (PCI-aware), discount campaigns, user management
    (Keycloak-backed) and behavioural event ingestion.

    Authentication is performed via Keycloak-issued JWT bearer tokens.
    Public endpoints (catalog browsing, registration, Stripe webhook)
    are explicitly whitelisted; every other endpoint requires a
    valid Bearer token and may further restrict to USER or ADMIN roles.
  contact:
    name: Cybertech Team
    email: dev@cybertech.local
  license:
    name: Proprietary
    url: https://cybertech.local/license
  version: "1.0"
servers:
- url: http://localhost:8081
  description: Local dev
- url: https://api.cybertech.local
  description: Production
security:
- keycloak: []
...
paths:
...
  /api/v1/services/review/create:
    post:
      tags:
      - Review
      summary: Create a new Review
      description: Registers a new Review in the system. Requires authenticated USER
        role.
      operationId: createReview
      parameters:
      - name: X-API-VERSION
        in: header
        schema:
          type: string
          default: 1.0.0
          enum:
          - "1.0"
      requestBody:
        description: Review data for creation. All fields are mandatory
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewCreateRequestDto"
        required: true
      responses:
        "201":
          description: Review created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewResponseDto"
        "400":
          description: Invalid input data / Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "401":
          description: Unauthorized - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: Forbidden - caller does not own the order being reviewed or
            comment moderation rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "404":
          description: Product / Order / User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "500":
          description: Internal server error during Review creation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
      security:
      - keycloak: []
  /api/v1/services/product/search:
    post:
      tags:
      - Product
      summary: Search products
      description: "Searches products using filters such as name, category, brand,\
        \ price range, attributes, etc."
      operationId: searchProducts
      parameters:
      - name: X-API-VERSION
        in: header
        schema:
          type: string
          default: 1.0.0
          enum:
          - "1.0"
      requestBody:
        description: Search criteria for filtering products
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductSearchRequestDto"
        required: true
      responses:
        "200":
          description: List of products matching the search criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductResponseDto"
        "400":
          description: Invalid search criteria
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
...
  /api/v1/services/review/update/{reviewUuid}:
    patch:
      tags:
      - Review
      summary: Update an existing Review by UUID
      description: Updates an existing review's details based on their unique UUID.
        Fields not provided will not be updated. Requires authenticated USER role
        and ownership of the review.
      operationId: updateReview
      parameters:
      - name: X-API-VERSION
        in: header
        schema:
          type: string
          default: 1.0.0
          enum:
          - "1.0"
      requestBody:
        description: Review data for update. Only provide fields that need to be changed.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewUpdateRequestDto"
        required: true
      responses:
        "200":
          description: Review updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewResponseDto"
        "400":
          description: Invalid input data / Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "401":
          description: Unauthorized - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: Operation forbidden - caller is not the author of the review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "404":
          description: Review not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "500":
          description: Internal server error during review update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
      security:
      - keycloak: []
...
  /api/v1/services/review/reviewable:
    get:
      tags:
      - Review
      summary: List the products the caller can review
      description: "Products the authenticated user has bought (order PAID, SHIPPED\
        \ or DELIVERED) and not reviewed yet, with the order to reference in POST\
        \ /create."
      operationId: getReviewableProducts
      parameters:
      - name: X-API-VERSION
        in: header
        schema:
          type: string
          default: 1.0.0
          enum:
          - "1.0"
      responses:
        "200":
          description: Reviewable products (possibly empty)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewableProductDto"
        "401":
          description: Unauthorized - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: Operation forbidden - USER role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
      security:
      - keycloak: []
  /api/v1/services/review/get/{reviewUuid}:
    get:
      tags:
      - Review
      summary: Request a Review by UUID
      description: Fetches a Review's details based on their unique UUID.
      operationId: getReviewByUuid
      parameters:
      - name: reviewUuid
        in: path
        description: UUID for searching a Review
        required: true
        schema:
          type: string
          format: uuid
      - name: X-API-VERSION
        in: header
        schema:
          type: string
          default: 1.0.0
          enum:
          - "1.0"
          - 1.0.0
      responses:
        "200":
          description: Review found successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewResponseDto"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: Operation forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "404":
          description: Review not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
  /api/v1/services/recommendations/my-recommendations:
    get:
      tags:
      - Recommendations
      summary: Get my recommendations
      description: "Returns up to n personalized product recommendations for the connected\
        \ user, sourced from Gorse with a fallback to newest-catalog then best-sellers\
        \ if Gorse is unavailable or has no history yet for this user."
      operationId: getMyRecommendations
      parameters:
      - name: "n"
        in: query
        description: Number of recommendations to return
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: X-API-VERSION
        in: header
        schema:
          type: string
          default: 1.0.0
          enum:
          - "1.0"
      responses:
        "200":
          description: Recommendations retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductResponseDto"
        "401":
          description: Unauthorized - User not logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
      security:
      - keycloak: []
  /api/v1/services/product/get/{productUuid}:
    get:
      tags:
      - Product
      summary: Request a Product by UUID
      description: Fetches a Product's details based on their unique UUID.
      operationId: getProductByUuid
      parameters:
      - name: productUuid
        in: path
        description: UUID for searching a Product
        required: true
        schema:
          type: string
          format: uuid
      - name: X-API-VERSION
        in: header
        schema:
          type: string
          default: 1.0.0
          enum:
          - "1.0"
          - 1.0.0
      responses:
        "200":
          description: Product found successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponseDto"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
  /api/v1/services/product/best-sellers:
    get:
      tags:
      - Product
      summary: Get best-selling products
      description: "Returns products ranked by total quantity sold, descending. Paginated."
      operationId: getBestSellers
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      - name: X-API-VERSION
        in: header
        schema:
          type: string
          default: 1.0.0
          enum:
          - "1.0"
      responses:
        "200":
          description: Page of best-selling products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponseDto"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
...
  /api/v1/services/review/delete/{reviewUuid}:
    delete:
      tags:
      - Review
      summary: Delete a Review by UUID
      description: Deletes a review based on their unique UUID. Requires authenticated
        USER role and ownership of the review.
      operationId: deleteReviewByUuid
      parameters:
      - name: reviewUuid
        in: path
        description: The UUID of the review to delete
        required: true
        schema:
          type: string
          format: uuid
      - name: X-API-VERSION
        in: header
        schema:
          type: string
          default: 1.0.0
          enum:
          - "1.0"
          - 1.0.0
      responses:
        "204":
          description: Review deleted successfully (No Content)
        "400":
          description: "Bad request (e.g., invalid UUID format)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "401":
          description: Unauthorized - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: Operation forbidden - caller is not the author of the review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "404":
          description: Review not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "500":
          description: Internal server error during review deletion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
      security:
      - keycloak: []
...
components:
  schemas:
...
    ErrorResponseDto:
      type: object
      properties:
        message:
          type: string
        httpStatusCode:
          type: integer
          format: int32
        errorCodeType:
          type: string
          enum:
          - TECHNICAL
          - FUNCTIONAL
    ProductResponseDto:
      type: object
      properties:
        name:
          type: string
        uuid:
          type: string
        price:
          type: number
        brand:
          type: string
        category:
          type: string
        photoUrls:
          type: array
          items:
            type: string
        description:
          type: string
        attributes:
          type: object
          additionalProperties: {}
...
    ReviewCreateRequestDto:
      type: object
      properties:
        orderUuid:
          type: string
          format: uuid
        productUuid:
          type: string
          format: uuid
        rating:
          type: integer
          format: int32
          maximum: 5
          minimum: 1
        comment:
          type: string
          maxLength: 1000
          minLength: 0
      required:
      - orderUuid
      - productUuid
      - rating
    ReviewResponseDto:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        userUuid:
          type: string
          format: uuid
        productUuid:
          type: string
          format: uuid
        productName:
          type: string
        rating:
          type: integer
          format: int32
        comment:
          type: string
    ProductSearchRequestDto:
      type: object
      properties:
        keyword:
          type: string
        priceMin:
          type: number
          format: double
          minimum: 100
        priceMax:
          type: number
          format: double
          maximum: 1000000
        category:
          type: string
        brands:
          type: array
          items:
            type: string
            enum:
            - ASUS
            - HP
            - LENOVO
            - DELL
            - AORUS
            - MSI
            - ACER
            - TOSHIBA
            - ALIENWARE
            - MICROSOFT
            - ALTYK
            - GIGABYTE
            - SAMSUNG
        attributes:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        numericRanges:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/RangeFilter"
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
    RangeFilter:
      type: object
      properties:
        min:
          type: number
          format: double
        max:
          type: number
          format: double
...
    ReviewUpdateRequestDto:
      type: object
      properties:
        reviewUuid:
          type: string
          format: uuid
        rating:
          type: integer
          format: int32
          maximum: 5
          minimum: 1
        comment:
          type: string
          maxLength: 1000
          minLength: 0
      required:
      - reviewUuid
...
    ReviewableProductDto:
      type: object
      properties:
        productUuid:
          type: string
          format: uuid
        orderUuid:
          type: string
          format: uuid
        productName:
          type: string
        orderDate:
          type: string
          format: date
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
...
  securitySchemes:
    keycloak:
      type: http
      description: "Keycloak-issued JWT access token. Obtain one from the realm token\
        \ endpoint and pass it as the Authorization header: 'Bearer <token>'."
      scheme: bearer
      bearerFormat: JWT