# src/main/java/com/novatech/cybertech/api/controllers/implementation/ProductSearchController.java
package com.novatech.cybertech.api.controllers.implementation;

import com.novatech.cybertech.api.controllers.spec.ProductSearchApiSpec;
import com.novatech.cybertech.dto.request.search.ProductSearchRequestDto;
import com.novatech.cybertech.dto.response.product.ProductResponseDto;
import com.novatech.cybertech.services.core.ProductManagementService;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.data.web.PageableDefault;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.UUID;

import static com.novatech.cybertech.constants.CyberTechAppConstants.*;
import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;

@RestController
@RequiredArgsConstructor
@RequestMapping(version = APP_API_VERSION, value = PRODUCT_CRUD_CONTROLLER_BASE_PATH)
public class ProductSearchController implements ProductSearchApiSpec {

    // Inject service interface instead of concrete impl per project convention
    private final ProductManagementService productService;

    @Override
    @GetMapping(value = "/get/{productUuid}", produces = APPLICATION_JSON_VALUE)
    public ResponseEntity<ProductResponseDto> getProductByUuid(@PathVariable("productUuid") UUID productUuid) {
        // Public, anonymous endpoint (whitelisted under /api/v1/services/product/** in
        // SecurityConfig#PUBLIC_URLS) — no caller identity exists to pass here.
        return ResponseEntity.status(HttpStatus.OK).body(productService.getByUUID(productUuid, null));
    }

    @Override
    // Removed redundant alias variable; @Override added to enforce ApiSpec contract at compile-time
    @PostMapping(value = "/search", consumes = APPLICATION_JSON_VALUE, produces = APPLICATION_JSON_VALUE)
    public Page<ProductResponseDto> searchProducts(@Valid @RequestBody final ProductSearchRequestDto productSearchRequestDto) {
        return productService.searchProducts(productSearchRequestDto);
    }

    @Override
    @GetMapping(value = "/best-sellers", produces = APPLICATION_JSON_VALUE)
    public ResponseEntity<Page<ProductResponseDto>> getBestSellers(
            @PageableDefault(size = DEFAULT_PAGE_SIZE_BEST_SELLERS) final Pageable pageable
    ) {
        return ResponseEntity.ok(productService.getBestSellers(pageable));
    }
}